docker/rescale-images/requirements.txt
pillow
pillow-avif-plugin
